Feature/156(로그인한 유저에 한한여 상품 등록/수정/삭제 가능하도록 기능 구현) (#187)
* feat: 로그인한 유저에 한한여 상품 등록/수정/삭제 가능하도록 기능 구현
-verifyUser를 통해 해당 상품을 수정 또는 삭제하려는 유저가 해당 상품을 등록한 유저인지 확인

* feat: 프로필 등록해야 상품 등록이 가능하도록 기능 구현

+++ backend/src/main/java/woorifisa/goodfriends/backend/product/presentation/ProductController.java
@@ -4,6 +4,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+import woorifisa.goodfriends.backend.auth.dto.LoginUser;
+import woorifisa.goodfriends.backend.auth.presentation.AuthenticationPrincipal;
 import woorifisa.goodfriends.backend.product.application.ProductService;
 import woorifisa.goodfriends.backend.product.dto.request.ProductSaveRequest;
 import woorifisa.goodfriends.backend.product.dto.request.ProductUpdateRequest;
@@ -11,10 +13,11 @@ import woorifisa.goodfriends.backend.product.dto.response.ProductSaveResponse;
 import woorifisa.goodfriends.backend.product.dto.response.ProductViewAllResponse;
 import woorifisa.goodfriends.backend.product.dto.response.ProductUpdateResponse;
 import woorifisa.goodfriends.backend.product.dto.response.ProductViewOneResponse;
+import woorifisa.goodfriends.backend.user.application.UserService;
+import woorifisa.goodfriends.backend.user.dto.response.UserResponse;
 
 import java.io.IOException;
 import java.net.MalformedURLException;
-import java.net.URI;
 import java.util.List;
 
 @RequestMapping("/api/products")
@@ -23,49 +26,73 @@ public class ProductController {
 
     private final ProductService productService;
 
-    public ProductController(ProductService productService) {
+    public ProductController(ProductService productService, UserService userService) {
         this.productService = productService;
     }
 
-    @PostMapping("/{userId}")
-    public ResponseEntity<Long> saveProduct(@PathVariable Long userId,
-                                                           @RequestPart ProductSaveRequest request,
-                                                           @RequestPart List<MultipartFile> multipartFiles) throws IOException {
-        ProductSaveRequest productSaveRequest = new ProductSaveRequest(request.getTitle(), request.getProductCategory(),request.getDescription(), request.getSellPrice(), multipartFiles);
-        ProductSaveResponse response = productService.saveProduct(userId, productSaveRequest);
-        return ResponseEntity.status(HttpStatus.CREATED).body(response.getId());
+    @PostMapping
+    public ResponseEntity<Long> saveProduct(@AuthenticationPrincipal final LoginUser loginUser,
+                                            @RequestPart ProductSaveRequest request,
+                                            @RequestPart List<MultipartFile> multipartFiles) throws IOException {
+        //프로필 등록해야 상품 등록 가능하도록
+        if(!productService.existProfile(loginUser.getId())) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN); // 403
+        }
+        else{
+            ProductSaveRequest productSaveRequest = new ProductSaveRequest(request.getTitle(), request.getProductCategory(),request.getDescription(), request.getSellPrice(), multipartFiles);
+            ProductSaveResponse response = productService.saveProduct(loginUser.getId(), productSaveRequest);
+            return ResponseEntity.status(HttpStatus.CREATED).body(response.getId()); // 201
+        }
     }
 
     @GetMapping("/view")
     public ResponseEntity<List<ProductViewAllResponse>> viewAllProduct() {
         List<ProductViewAllResponse> responses = productService.viewAllProduct();
-        return ResponseEntity.ok().body(responses);
+        return ResponseEntity.ok().body(responses); // 200
     }
 
     @GetMapping("/view/{productId}")
     public ResponseEntity<ProductViewOneResponse> viewOneProduct(@PathVariable Long productId) {
         ProductViewOneResponse response = productService.viewOneProduct(productId);
-        return ResponseEntity.ok().body(response);
+        return ResponseEntity.ok().body(response); // 200
     }
 
     @GetMapping("/edit/{productId}")
-    public ResponseEntity<ProductUpdateResponse> showSelectedProduct(@PathVariable Long productId){
-        ProductUpdateResponse response = productService.showSelectedProduct(productId);
-        return ResponseEntity.ok().body(response);
+    public ResponseEntity<ProductUpdateResponse> showSelectedProduct(@AuthenticationPrincipal final LoginUser loginUser,
+                                                                     @PathVariable Long productId){
+        if(productService.verifyUser(loginUser.getId(), productId)) {
+            ProductUpdateResponse response = productService.showSelectedProduct(productId);
+            return ResponseEntity.ok().body(response); // 200
+        }
+        else {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN); // 403
+        }
     }
 
     @PutMapping("/edit/{productId}")
-    public ResponseEntity<Void> updateProduct(@PathVariable Long productId,
-                                                               @RequestPart ProductUpdateRequest request,
-                                                               @RequestPart List<MultipartFile> multipartFiles) throws IOException {
-        ProductUpdateRequest productUpdateRequest = new ProductUpdateRequest(request.getTitle(), request.getProductCategory(), request.getDescription(), request.getSellPrice(), multipartFiles);
-        ProductUpdateResponse response = productService.updateProduct(productUpdateRequest, productId);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    public ResponseEntity<Void> updateProduct(@AuthenticationPrincipal final LoginUser loginUser,
+                                              @PathVariable Long productId,
+                                              @RequestPart ProductUpdateRequest request,
+                                              @RequestPart List<MultipartFile> multipartFiles) throws IOException {
+        if(productService.verifyUser(loginUser.getId(), productId)) {
+            ProductUpdateRequest productUpdateRequest = new ProductUpdateRequest(request.getTitle(), request.getProductCategory(), request.getDescription(), request.getSellPrice(), multipartFiles);
+            productService.updateProduct(productUpdateRequest, productId);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT); // 204
+        }
+        else {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN); // 403
+        }
     }
 
     @DeleteMapping("/delete/{productId}")
-    public ResponseEntity<Void> deleteProduct(@PathVariable Long productId) throws MalformedURLException {
-        productService.deleteById(productId);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    public ResponseEntity<Void> deleteProduct(@AuthenticationPrincipal final LoginUser loginUser,
+                                              @PathVariable Long productId) throws MalformedURLException {
+        if(productService.verifyUser(loginUser.getId(), productId)) {
+            productService.deleteById(productId);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT); // 204
+        }
+        else {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN); // 403
+        }
     }
 }
